requirements.txt
absl-py==2.0.0
beautifulsoup4==4.12.2
brotlipy==0.7.0
cachetools==5.3.2
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi @ file:///croot/cffi_1670423208954/work
chardet==5.2.0
charset-normalizer==3.3.2
cryptography @ file:///croot/cryptography_1677533068310/work
cycler==0.11.0
Cython @ file:///croot/cython_1676568029361/work
easydict==1.11
faiss-gpu==1.7.2
filelock==3.12.2
fonttools==4.38.0
fsspec==2023.1.0
future==1.0.0
gdown==4.7.1
google-auth==2.23.4
google-auth-oauthlib==0.4.6
grpcio==1.59.3
h5py==3.8.0
huggingface-hub==0.16.4
idna==3.6
importlib-metadata==6.7.0
joblib==1.3.2
kiwisolver==1.4.5
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.5.3
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626179032232/work
mkl-service==2.4.0
numpy @ file:///opt/conda/conda-bld/numpy_and_numpy_base_1653915516269/work
oauthlib==3.2.2
opencv-python==4.8.1.78
packaging==23.2
Pillow==9.4.0
prettytable==3.7.0
protobuf==3.20.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyOpenSSL @ file:///croot/pyopenssl_1677607685877/work
pyparsing==3.1.1
PySocks @ file:///tmp/build/80754af9/pysocks_1594394576006/work
python-dateutil==2.8.2
pytorch-ignite==0.1.2
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.3
scikit-learn==1.0.2
scipy @ file:///home/lucky7/scipy-1.2.0-cp37-cp37m-manylinux1_x86_64.whl
six @ file:///tmp/build/80754af9/six_1644875935023/work
soupsieve==2.4.1
tabulate==0.9.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==2.3.0
threadpoolctl==3.1.0
timm==0.9.12
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tqdm==4.66.1
typing_extensions==4.7.1
urllib3==2.0.7
wcwidth==0.2.12
Werkzeug==2.2.3
yacs==0.1.8
zipp==3.15.0
